# Improve speed of GROMOS11 reader