--- a/package/AUTHORS
+++ b/package/AUTHORS
@@ -258,6 +258,7 @@ Chronological list of authors
   - Abdulrahman Elbanna
   - Tulga-Erdene Sodjargal
   - Gareth Elliott
+  - Marc Schuh
 
 
 External code