--- a/package/CHANGELOG
+++ b/package/CHANGELOG
@@ -38,6 +38,7 @@ Fixes
    directly passing them. (Issue #3520, PR #5006)
 
 Enhancements
+ * Improve speed of GROMOS11 (TRC) reader (Issue #5079)
  * Improve parsing of topology information from LAMMPS dump files to allow
    reading of mass, charge and element attributes. (Issue #3449, PR #4995)
  * Added timestep support for XDR writers and readers (Issue #4905, PR #4908)