--- a/testsuite/MDAnalysisTests/datafiles.py
+++ b/testsuite/MDAnalysisTests/datafiles.py
@@ -336,6 +336,7 @@ __all__ = [
     "GSD_long",
     "TRC_PDB_VAC",
     "TRC_TRAJ1_VAC",
+    "TRC_TRAJ1_VAC_WHITESPACE",  # contains inconsistent trailing whitespace in POSITIONRED
     "TRC_TRAJ2_VAC",  # 2x 3 frames of vacuum trajectory from GROMOS11 tutorial
     "TRC_CLUSTER_VAC",  # three frames without TIMESTEP and GENBOX block but with unsupported POSITION block
     "TRC_TRICLINIC_SOLV",
@@ -839,6 +840,7 @@ GSD_long = (_data_ref / "example_longer.gsd").as_posix()
 
 TRC_PDB_VAC = (_data_ref / "gromos11/gromos11_traj_vac.pdb.gz").as_posix()
 TRC_TRAJ1_VAC = (_data_ref / "gromos11/gromos11_traj_vac_1.trc.gz").as_posix()
+TRC_TRAJ1_VAC_WHITESPACE = (_data_ref / "gromos11/gromos11_traj_vac_1_whitespace.trc.gz").as_posix()
 TRC_TRAJ2_VAC = (_data_ref / "gromos11/gromos11_traj_vac_2.trc.gz").as_posix()
 TRC_PDB_SOLV = (_data_ref / "gromos11/gromos11_traj_solv.pdb.gz").as_posix()
 TRC_TRAJ_SOLV = (_data_ref / "gromos11/gromos11_traj_solv.trc.gz").as_posix()