… testsuite/MDAnalysisTests/data/gromos11/gromos11_traj_vac_1_extra_pos.trc.gz
… testsuite/MDAnalysisTests/data/gromos11/gromos11_traj_vac_1_missing_pos.trc.gz
… testsuite/MDAnalysisTests/data/gromos11/gromos11_traj_vac_1_whitespace.trc.gz